Updated and Moved question add, update and delete queries to questions/[...params] and passed selectedList to api calls as param
